refactor: Extract getHierarchyRoot
