+++ src/charts/BubbleChart.ts
@@ -1,4 +1,3 @@
-import { hierarchy, pack } from "d3-hierarchy";
 import { Datum } from ".";
 import { ColorMap } from "../colors";
 import { Svg } from "../components";
@@ -7,12 +6,11 @@ import { Dimensions, ResolvedDimensions } from "../dims";
 import { BaseMax, BubbleInputStep, InputGroupValue, TextDims } from "../types";
 import { FONT_SIZE, getTextColor, max } from "../utils";
 import * as Chart from "./Chart";
-import { HierarchyRoot } from "./types";
 import {
-  PADDING,
   STROKE_WIDTH,
   getBaseMax,
   getGroupLabelStrokeWidth,
+  getHierarchyRoot,
   getRotate,
 } from "./utils";
 
@@ -70,7 +68,7 @@ export const getters = (
     colorMap,
     cartoonize,
   } = props;
-  const root = getRoot({ groups, size: maxValue.k * size });
+  const root = getHierarchyRoot({ groups, size: maxValue.k * size });
   const groupsGetters: Chart.Getter[] = [];
   // If a custom maxValue was provided, we need to shift the bubbles to the center.
   const maxValueShift = maxValue.kc * size * 0.5;
@@ -191,27 +189,4 @@ export const getters = (
   }
 
   return groupsGetters;
-};
-
-const getRoot = ({
-  groups,
-  size,
-}: {
-  groups: InputGroupValue[];
-  size: number;
-}): HierarchyRoot => {
-  const root = hierarchy({
-    children: groups.map((d) => ({
-      key: d.key,
-      opacity: d.opacity,
-      children: d.data,
-    })),
-  }).sum((d) => Math.max(0, (d as any).value));
-  const descendants = root.descendants();
-  const leaves = descendants.filter((d) => !d.children);
-  leaves.forEach((d: any, i) => (d.index = i));
-  root.sort((a: any, b: any) => a.index - b.index);
-  pack().size([size, size]).padding(PADDING)(root as any);
-
-  return root as any as HierarchyRoot;
 };
